Move Recommendations part down
This part was intended as a relatively brief overview of how to proceed,
in a manner similar to putting the conclusion at the top. It was clear
from the presentation that talking about "Oracle JDK" and "Oracle
OpenJDK" without attempting to establish those terms was confusing.
Perhaps talking about the distributions beforehand makes this easier.